--- a/packages/back-end/scripts/stats_server.py
+++ b/packages/back-end/scripts/stats_server.py
@@ -7,24 +7,48 @@ from gbstats.gbstats import process_multiple_experiment_results
 for line in sys.stdin:
     start = time.time()
 
-    # Read from stdin
-    input = json.loads(line, strict=False)
-
-    id = input["id"]
-
-    # cast asdict because dataclasses are not serializable
-    results = [asdict(analysis) for analysis in process_multiple_experiment_results(input["data"])]
+    # Read from stdin and parse JSON
     try:
+        input = json.loads(line, strict=False)
+    except json.JSONDecodeError as e:
+        sys.stderr.write(f"Invalid JSON input: {str(e)}\n")
+        sys.stderr.flush()
+        continue
+    except Exception as e:
+        sys.stderr.write(f"Unexpected error parsing input: {str(e)}\n")
+        sys.stderr.flush()
+        continue
+
+    # Extract required fields
+    try:
+        id = input["id"]
+        data = input["data"]
+    except KeyError as e:
+        sys.stderr.write(f"Missing required field: {str(e)}\n")
+        sys.stderr.flush()
+        continue
+    except TypeError as e:
+        sys.stderr.write(f"Input is not a valid object: {str(e)}\n")
+        sys.stderr.flush()
+        continue
+    except Exception as e:
+        sys.stderr.write(f"Error extracting fields from input: {str(e)}\n")
+        sys.stderr.flush()
+        continue
+
+    # Process experiment results
+    try:
+        results = [asdict(analysis) for analysis in process_multiple_experiment_results(data)]
         sys.stdout.write(json.dumps({
             'id': id,
             'results': results,
             'time': time.time() - start
         }, allow_nan=False) + "\n")
+        sys.stdout.flush()
     except Exception as e:
         sys.stdout.write(json.dumps({
             'id': id,
-            'error': str(e),
+            'error': f"Processing error: {str(e)}",
             'time': time.time() - start
         }, allow_nan=False) + "\n")
-
-    sys.stdout.flush()
+        sys.stdout.flush()