Currently, whenever we call the stats engine from Node, we generate a temporary `.py` file, write to the file system, and then call it in a new shell.  This adds an incredible amount of overhead.

This PR instead creates a pool of long-running python processes and uses stdin/stdout to pass data between Node and Python.

Some quick benchmarks on my laptop:

| Metric | Before | After |
| ---------|-----------------|------ |
| Python time | 673ms | 700ms |
| Node.js overhead | 1714ms | 5ms |
| Total time | 2387ms | 705ms |
| Avg CPU | 18% | 9% |

This has a huge impact on scalability.  Before, a rate of 2.4 calls per second was enough to cause a 100% CPU load.  After this change, it takes 16 calls per second to do the same.

Importantly, our job queue runs up to 5 experiment refreshes in parallel, so it's not uncommon to see periods of >100% CPU load in production.  This change should eliminate that issue.

TODO:
- [ ] Test within docker build
- [ ] Test error handling (at all levels)
- [ ] Test edge cases
- [ ] Test pool / locking behavior with many parallel calls